refactor: move ex_retab() to indent.c
